+++ barcodes/Services/DataMatrixEncoder+Encoding.swift
@@ -6,8 +6,6 @@ extension DataMatrixEncoder {
     // MARK: - ASCII Encoder
 
     struct ASCIIEncoder {
-        let encodingMode = EncodingMode.ascii
-
         func encode(_ ctx: EncoderContext) {
             let digitCount = consecutiveDigitCount(
                 ctx.message, startpos: ctx.pos
@@ -164,8 +162,6 @@ extension DataMatrixEncoder {
     // MARK: - C40 Encoder
 
     struct C40Encoder {
-        let encodingMode = EncodingMode.c40
-
         func encode(_ ctx: EncoderContext) {
             var buffer: [UInt8] = []
             while ctx.hasMoreCharacters {
@@ -221,8 +217,6 @@ extension DataMatrixEncoder {
     // MARK: - Text Encoder
 
     struct TextEncoder {
-        let encodingMode = EncodingMode.text
-
         func encode(_ ctx: EncoderContext) {
             var buffer: [UInt8] = []
             while ctx.hasMoreCharacters {